\name{caiman_crocodilus}
\docType{data}
\alias{caiman_crocodilus}
\title{Subset of Caiman cocodrilus records for the year 2020}
\description{ This is a example dataset generated by [get_specie_records()]
}
\usage{caiman_crocodilus}
\format{A data.frame containing 100 observations.}
\source{GBIF API}
\keyword{data}
