% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\docType{data}
\name{departamentos}
\alias{departamentos}
\title{Colombian departament boundaires}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 33 rows and 2 columns.
}
\usage{
departamentos
}
\description{
Colombian departament boundaires
}
\references{
Moreno D (2015). \emph{colmaps: Colombian maps}. R package version 0.0.0.9515.
}
\seealso{
\link[colmaps:departamentos]{colmaps::departamentos}
}
\keyword{data}
