% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling_functions.R
\name{filter_if_match}
\alias{filter_if_match}
\title{Filter a data.frame using regex}
\usage{
filter_if_match(df, column, terms, exclude = FALSE)
}
\arguments{
\item{df}{A data.frame.}

\item{column}{A string with the column name to filter.}

\item{terms}{A vector with a pattern to match or regular expressions.}

\item{exclude}{A boolean,
\itemize{
\item \code{TRUE}: Keep values that does not match terms.
\item \code{FALSE}(the default): Keep just those values whose match terms.
}}
}
\value{
A data.frame
}
\description{
This function is a wrapper of \code{\link[dplyr:filter]{dplyr::filter()}} and \code{\link[stringr:str_detect]{stringr::str_detect()}} that makes partial matching easier.
}
\examples{
filter_if_match(iris,"Species",terms = "ica", exclude = F)
filter_if_match(iris,"Species",terms = "^s\\\\w+a$", exclude = F)

}
