% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{map_colombia}
\alias{map_colombia}
\title{Create a colombia map with specie records}
\usage{
map_colombia(
  df,
  coords = c("decimalLongitude", "decimalLatitude"),
  legend.show = FALSE,
  column
)
}
\arguments{
\item{df}{A data.frame with coordinates.}

\item{coords}{A character vector with the column names that store record coordinates. the default asums coords = c("decimalLongitude", "decimalLatitude") using a data.frame generated by \code{\link[=get_specie_records]{get_specie_records()}}}

\item{legend.show}{A Boolean. It is a wrapper of legend.show from \code{\link[tmap:tm_layout]{tmap::tm_layout()}}.}

\item{column}{A character vector. It is the name column you want to plot.}
}
\value{
A tmap object.
}
\description{
Create a colombia map with specie records
}
\examples{
map_colombia(caiman_crocodilus,column = "scientificName",legend.show = F)
}
