% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{map_depto}
\alias{map_depto}
\title{Create a map for a Colombian department with specie records.}
\usage{
map_depto(
  df,
  depto,
  coords = c("decimalLongitude", "decimalLatitude"),
  legend.show = FALSE,
  column = "scientificName"
)
}
\arguments{
\item{df}{A data.frame with coordinates.}

\item{depto}{A character vector. It must be a single string with valid name from departamentos$depto.}

\item{coords}{A character vector with the column names that store record coordinates. the default asums coords = c("decimalLongitude", "decimalLatitude") using a data.frame generated by \code{\link[=get_specie_records]{get_specie_records()}}}

\item{legend.show}{logicl. It is a wrapper of legend.show from \code{\link[tmap:tm_layout]{tmap::tm_layout()}}.}

\item{column}{A character vector. It is the name column you want to plot.}
}
\value{
A tmap object.
}
\description{
Create a map for a Colombian department with specie records.
}
\examples{
map_depto(caiman_crocodilus, depto = "La Guajira", column = "scientificName",legend.show = FALSE)
}
