% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\name{map_mun}
\alias{map_mun}
\title{Create a map for a Colombian municipality with specie records}
\usage{
map_mun(
  df,
  municipio,
  coords = c("decimalLongitude", "decimalLatitude"),
  legend.show = FALSE,
  column = "scientificName"
)
}
\arguments{
\item{df}{A data.frame with coordinates.}

\item{municipio}{A character vector. It must be a single string with valid name from municipios$municipio.}

\item{coords}{A character vector with the column names that store record coordinates. the default asums coords = c("decimalLongitude", "decimalLatitude") using a data.frame generated by \code{\link[=get_specie_records]{get_specie_records()}}}

\item{legend.show}{logical. It is a wrapper of legend.show from \code{\link[tmap:tm_layout]{tmap::tm_layout()}}.}

\item{column}{A character vector. It is the name column you want to plot.}
}
\value{
A tmap object.
}
\description{
Create a map for a Colombian municipality with specie records
}
\examples{
map_mun(caiman_crocodilus, mun = "Barrancas", column = "scientificName",legend.show = FALSE)
}
