% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_functions.R
\docType{data}
\name{municipios}
\alias{municipios}
\title{Colombian municipality boundaries}
\format{
An object of class \code{SpatialPolygonsDataFrame} with 1122 rows and 4 columns.
}
\usage{
municipios
}
\description{
Colombian municipality boundaries
}
\references{
Moreno D (2015). \emph{colmaps: Colombian maps}. R package version 0.0.0.9515.
}
\seealso{
\link[colmaps:municipios]{colmaps::municipios}
}
\keyword{data}
