% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSiB_API_functions.R
\name{print_counter}
\alias{print_counter}
\title{Print a simple counter}
\usage{
print_counter(counter, offset)
}
\arguments{
\item{counter}{An integer, The counter's startpoint.}

\item{offset}{An integer, the page number from SiB API.}
}
\value{
An integer.
}
\description{
Simple counter printer used in \code{\link[=get_specie_records]{get_specie_records()}} to give information to the user.
}
