% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_wrangling_functions.R
\name{relevant_vars}
\alias{relevant_vars}
\title{Select relevant variables from GBIF data.frame}
\usage{
relevant_vars(df, drop.coor.na = TRUE)
}
\arguments{
\item{df}{A data.frame generated by \code{\link[=get_specie_records]{get_specie_records()}}.}

\item{drop.coor.na}{Logical TRUE (The deault) drop NA values from "decimalLongitude", "decimalLatitude" columns. FALSE keep NAs.}
}
\value{
A data.frame.
}
\description{
Select most revant columns from a data.frame generated by \code{\link[=get_specie_records]{get_specie_records()}}, . Also this function drop NAs from  "decimalLongitude", "decimalLatitude" columns.
}
\examples{
relevant_vars(caiman_crocodilus)
}
